% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_build.R
\name{tab_build}
\alias{tab_build}
\title{Créer un tableau croise}
\usage{
tab_build(
  data,
  var_rows,
  var_cols = NULL,
  var_stat = NULL,
  stat = "count",
  var_w = NULL,
  probs = NULL,
  lab_total = "Ensemble"
)
}
\arguments{
\item{data}{table (data.frame ou tibble) contenant les données}

\item{var_rows}{variables lignes (caractère ou facteur)}

\item{var_cols}{variables colonnes (caractère ou facteur)}

\item{var_stat}{variable d'intérêt sur laquelle calculer la statististique (numérique)}

\item{stat}{statistique à calculer
Ce paramètre peut prendre les valeurs suivantes :
\itemize{
\item{\code{"mean"} : moyenne}
\item{\code{"min"} : minimum}
\item{\code{"max"} : maximum}
\item{\code{"median"} : médiane}
\item{\code{"sum"} : somme}
\item{\code{"count"} : effectifs - dans ce cas, laisser le parametre var_stat non renseigné}
\item{\code{"quantile"} : quantile}
\item{\code{"weighted_mean"} : moyenne pondérée}
\item{\code{"weighted_sum"} : somme pondérée}
}
Pour les statistiques pondérées, renseigner le paramètre var_w avec la variable de pondération à utiliser pour les calculs}

\item{var_w}{variable de pondération éventuelle}

\item{probs}{quantile à calculer si la statistique demandée est "quantile" - 0.5 pour la médiane, 0.1 pour le premier décile...}

\item{lab_total}{libellé pour les colonnes de total}
}
\value{
tableau croisé
}
\description{
tab_build() permet construire un tableau croisé : elle calcule une statistique pour plusieurs croisements de
variables, en intégrant automatiquement les différentes marges.
}
\examples{
library(dplyr)

prenoms |>
  filter(annais == "2020" &
             lib_reg \%in\% c("Occitanie", "Pays de la Loire")) |>
  tab_build(var_rows = lib_reg,
            var_cols = c(prenom, sexe),
            var_stat = nombre,
            lab_total = "Total",
            stat = "sum")
}
