% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_mask.R
\name{tab_mask}
\alias{tab_mask}
\title{Masquer les cellules d'un tableau}
\usage{
tab_mask(
  tab,
  threshold_count = NULL,
  threshold_sum = NULL,
  mask_total = "",
  verbose = TRUE
)
}
\arguments{
\item{tab}{tableau croisé issu de tab_build}

\item{threshold_count}{nombre de lignes de la table initiale en-deça duquel la valeur est masquée (exclusif) - par défaut : NULL}

\item{threshold_sum}{à renseigner pour stat = sum, valeur de la cellule en-deça de laquelle la valeur est masquée (exclusif) - par défaut : NULL}

\item{mask_total}{masquage ou non des totaux - par défaut : aucun
Ce paramètre peut prendre les valeurs suivantes :
\itemize{
\item{\code{"cols"} : masquage des totaux en colonnes}
\item{\code{"rows"} : masquage des totaux en lignes}
\item{\code{c("rows", "cols")} : masquage de l'ensemble des totaux}
}}

\item{verbose}{affichage ou non du nombre de cellues masquées par la fonction - par défaut : TRUE}
}
\value{
tableau avec cellules blanchies
}
\description{
tab_mask() permet de masquer les cellules d'un tableau :
\itemize{
\item les cellules non-nulles inférieures à un seuil, en nombre de lignes ou en somme de la variable var_stat
\item les cellules correspondant aux marges
}
}
\examples{
library(dplyr)

prenoms |>
  filter(annais == "2020" &
       prenom \%in\% c("Charlie", "Camille") &
                lib_reg \%in\% c("Occitanie", "Pays de la Loire")) |>
  tab_build(var_rows = lib_reg,
            var_cols = c(prenom, sexe),
            var_stat = nombre,
            lab_total = "Total",
            stat = "sum") |>
  tab_mask(threshold_count = 5,
           threshold_sum = 100,
           mask_total = c("rows", "cols"),
           verbose = TRUE)
}
