% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_pct.R
\name{tab_pct}
\alias{tab_pct}
\title{Calculer des pourcentages en lignes ou en colonnes}
\usage{
tab_pct(tab, pct)
}
\arguments{
\item{tab}{tableau croisé issu de tab_build}

\item{pct}{type de pourcentage souhaité
\itemize{
\item{\code{"row"} : pourcentage en lignes}
\item{\code{"col"} : pourcentage en colonnes}
}}
}
\value{
tableau transformé en pourcentages lignes ou colonnes
}
\description{
à partir d'un tableau croisé contenant des effectifs ou des sommes,
tab_pct() calcule des pourcentages lignes ou colonnes.
}
\examples{
library(dplyr)

prenoms |>
  filter(annais == "2020" &
         prenom \%in\% c("Charlie", "Eden", "Camille")) |>
  tab_build(var_rows = lib_reg,
            var_cols = c(prenom, sexe),
            var_stat = nombre,
            lab_total = "Total",
            stat = "sum") |>
  tab_pct(pct = "col")
}
