% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_render.R
\name{tab_render}
\alias{tab_render}
\title{Afficher un tableau mis en forme au format kable}
\usage{
tab_render(tab, title = "", note = "", lab_rows = NULL, lab_cols = NULL)
}
\arguments{
\item{tab}{tableau croisé issu de tab_build}

\item{title}{titre - par défaut : "" (aucun)}

\item{note}{note de bas de page - par défaut : "" (aucun)}

\item{lab_rows}{nouveaux libellés pour les variables lignes - par défaut : NULL}

\item{lab_cols}{nouveaux libellés pour les modalités des variables colonnes - par défaut : NULL}
}
\value{
tableau mis en forme pour insertion html (kable)
}
\description{
tab_render() met en forme un tableau croisé au format html, pour le visualiser ou l'insérer dans
un rappport md.
}
\examples{
library(dplyr)

prenoms |>
filter(annais \%in\% c("1900", "2020") &
         prenom == "Camille") |>
  tab_build(var_rows = lib_reg,
            var_cols = c(annais, sexe),
            lab_total = "Ensemble",
            stat = "count") |>
  tab_render(title = "Nombre d'enfants prénommés Camille en 1900 et en 2020",
             note = "Source : Insee - base des prénoms",
             lab_rows = "Région",
             lab_cols = list(sexe = c("Garçons", "Filles")))
}
