% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_xlsx.R
\name{tab_xlsx}
\alias{tab_xlsx}
\title{Exporter un tableau dans un classeur xlsx}
\usage{
tab_xlsx(
  tab,
  path,
  sheet,
  title = "",
  note = "",
  lab_rows = NULL,
  lab_cols = NULL,
  open_wb = TRUE,
  update = TRUE
)
}
\arguments{
\item{tab}{tableau croisé issu de tab_build}

\item{path}{chemin dans lequel exporter le tableau}

\item{sheet}{nom de l'onglet dans le classeur - si l'onglet existe déjà, il sera mis à jour avec update = TRUE}

\item{title}{titre - par défaut : "" (aucun)}

\item{note}{note de bas de page - par défaut : "" (aucun)}

\item{lab_rows}{nouveaux libellés pour les variables lignes - par défaut : NULL}

\item{lab_cols}{nouveaux libellés pour les modalités des variables colonnes - par défaut : NULL}

\item{open_wb}{ouverture du classeur crée pour visualisation - par défaut : TRUE}

\item{update}{mise à jour de l'onglet s'il existe déjà dans le classeur - par défaut : TRUE}
}
\value{
classeur excel contenant les données exportées
}
\description{
tab_xlsx() permet d'enregistrer le tableau mis en forme dans un fichier .xlsx.
}
\examples{
\dontrun{
library(dplyr)

prenoms |>
  filter(annais == "2020" &
           lib_reg \%in\% c("Occitanie", "Pays de la Loire")) |>
  tab_build(var_rows = c(lib_reg, prenom),
            var_cols = sexe,
            var_stat = nombre,
            lab_total = "Total",
            stat = "sum") |>
  tab_xlsx(title = "Nombre de naissances avec des prénoms mixtes en 2020",
           note = "Source : Insee - base des prénoms",
           lab_cols = list(sexe = c("Garçons", "Filles")),
           lab_rows = c("Région", "Prénom"),
           path =  "U:/test_export.xlsx",
           sheet = "naissances_1")
}
}
